+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - Implement SSO authentication
 - Rewrite login page
 - Remove DEMO support
+- Fixed text imported formula incorrect percentage handling
 
 ### Version 12.4
 - Add dilutant in API formulas